Debug Menu
First I made a debug menu that is acessable by writing debug in to the main program. After that I made a debug mode checker function which just checks the array to see if the debug id is present.

Currently its not working becuase python must be doing something to the user input and messing up the way it is stored