School
First I made a new version because a new day. After that I copied the file in to final version. Then i finished the color picker byh resorting to a set menu of colors to set it to. The user inputted colours wouldnt work no matter what I tried, I think this is happening becuase oython must do something to input varibles which messes up the ansi. After that I made a checker that checks if you are in running an idle terminal. This is because idle messes with the CMD commands used with ANSI coloring and Console clearing. I then added a Debug check so that when Im debugging the program I dont have to go through the idle check menu. After that I started working on the user infomation. But before I managed to finish that the bell rang.