Temp File: Makeing Writing and Reseting
First I wanted to get the temp dir so I used pythons built in way of doing that. Ufter reading up on the documentation I decided to not use a temp file and instead make a txt in the temp dir that gets reset with every run.  Then i made the reset and code i found it intresting that f.write creates the file. When I was testing it in windows terminal that for some reason  it thinks its running in idle.

[CONTINUE NEXT COMMIT]