Customer Details
First I added some debug info mation into the already there code. After that I went through and added the array and append lines. I then rearaged the code to fit the array's structure. After that I tested it out and relised that it was asking for the user input before the menu was beign printed. After fixing that I added a array reset if function is ran again so that the data is not indexed incorrectly. Next I added a check that checked if the phone number was actually a number. While adding this to the debug menu I also made a print document function.