The basics, Fish input.
After adding the menu I made a function to add the fish to the order. The function gets the user's number input, the items in the list and the prices. First it asks for the amount of fish of the item you want. After that it checks if the input is an int. Next usign a for loop: it firsts check if the amount of the fish is at the max, if it isnt then it adds 1 to the amount of max fish, then it addes the info to the user order and the session txt.