Finishing touches to the add to order.
First I added the fish name to the input. and then also to the added item text. After that i made a range if statment for the fish menu. The range was simpler and more scaleable then just copy and pasting multiple times. I then added the cost var to function becuase i forgot about it before.