Finish()
To make the finish menu i first copied the normal menu code to a new function. Then i made it check wether you had entered any info into the lists, after that the printDualMenu function is used to print the user data lists. Next it checks the types of the order, If it is frozen then it takes away 1.05 for everything in the list and if it is for delivery then it adds $5.                                                                                                              Next it prints the name and phone number and if the previous check returned delivery then it prints the adress. After that it writes it to history and then asks the operator to place another order or quit. If the operator chooses to place another order then it resets the file, lists and vars.