Venv Setup
After messing around with venv I have worked it out. I first made the enviroment in Python/VIRTUAL. After that i added in paths to it in the program.  After that i setup a debug tag called venv. The reason it is debug and not main stream is because you have to set all the paths. I will work on a way to make it more flexable but for now it is debug. After that i added 2 entrys "Run a file in venv" and "Install a package".