Run current program in venv
To make a way for this to not loop for ever i had to make a way of it know if a process had started. To do this used a temp file. I was goign to use a sub process checker like i did with idle but with cmd it would be too unreliable.