Added indiviadual check
i added an indviaul check for how many fish did the user order. Now instead of having a max of 7 total fish, there is now a max of 82 (7 each). This was a simple array checker that checks how many fish in the current order array is eqaul to  the inputted fish, then it just adds the relevant number