Updates from school
First i added the individual fish checker from the old version to the main one. Next i setup the venv config files to work with school and home. After that I edited the the venv vars to automatically get the current working dirctory. I then updated the debug menu to include this. After that i made a extremly simple counter that counts the items in a an array. But that will be focused on in the next commit