Dynamic input for the menus.
First i removed the windows tag becuase it  wasnt used in the program. After that i added a return for the array counter function. I then put this function to use by making the exit option on most of the menus more dynamic. Right before it prints the menu it appends "Exit" to the menu's array, it then counts the array and returns the value where exit is. Then it checks user input for the value. That way I can append options to the menu under met states eg. if venv is active then add the options to install a package or run a file, even with two new options exit is still last. Lastly i added a exit function to the idle menu so that it would close when the process was closed