School
First moved all the virtual enviroment code into a function. Then i fixed the file checker, and i also fixed the data checker so that if the customer info hasnt been entered then it will put you on that menu. Then i also finished the chatbot options in the debug menu and thats pretty much it