Updating the chat bot
I just added the paths to the chat bot. Then i made a file checker that checks a file for a string and returns the line below it, this is used for custom venvs. After that I fixed the code that runs the chat bot to work with proper /s and \s because programs treat \ as an escape and it messes with some of the code. Then