Fixing the insta kill
for some reason when ever trying to run in the virtual enviroment it automaticly quits. After doing some testing it turns out it wasnt a problem with the script, rather the loction of python beign incorrect so i fixed that. Next i researched a bit and found a way to run a function from another python file. After a bit of work i got it to work. The only problem was that it would run the file first to setup and then the function, so when i was testing this the file would run all the start up code i had. So to fix this I added a string to data.txt. When this string is detected then all the startup code wont be run, then once the external program has finished running the file is cleared of that tag.