Fixing the scoop error.
Currently there is an error where any input is treated as an amount of scoops. After an hour of trial and error i found that the problem is somewhere to do with fish amount. While searching for the previous error i also notice that they ran the chips() instead of fish() so i fiex that. WRITE THIS OUT: ADDED THE PRINT HIDER, FIXING FISH MENU