Data deletion
To fit in with the privacy Implication I had to add the data deltion function. To do this i first needed to make two new functions. The first function checks for the name on the same line as Customer Name:  and returns the line. The second one gos to the line that was returned by the previous function and counts each line until it reaches the specified string. Finnaly the deletePortion function delets the lrange of lines in (function 1, function 2)