changes
TIME WISE THIS IS FROM PREVIOUS COMMIT TO THIS ONE

I added a string of code that titles the cmd window "Freddy's Fast Fish", i know it is small but it counts. I then went through all the error displays and added what they inputted to the error.  I then changed the formatting so that it would support numbers up to 999 and costs worth 5 chars (99.99) wont break the theming (OCD). I then changed the chip check to be flexible so that the menu could be edited more easily. I also added some more comment code,. Afterwards I added the option to the debug menu to find which lines in the history file a customers details are on, and after that you can choose to print their data. I then fixed all the quit() functions to properly  write to history so that there are no  EOF errors. And i reworte reset() so that it worked as intented